requirements.txt
dify_plugin>=0.1.0,<0.2.0
openai>=1.13.0
requests>=2.31.0
Pillow>=10.0.0
tiktoken>=0.7.0
